dashboard/Portofolio/portfolio_analyzer_api_v2/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.8.0
httpx>=0.27.0
pytest>=8.0.0
openai>=1.0.0
python-dotenv>=1.0.0
